+++ zcash_client_backend/src/data_api/error.rs
@@ -6,9 +6,8 @@ use zcash_primitives::{
     consensus::BlockHeight,
     sapling::Node,
     transaction::{builder, components::amount::Amount, TxId},
+    zip32::AccountId,
 };
-
-use crate::wallet::AccountId;
 
 #[derive(Debug)]
 pub enum ChainInvalid {
@@ -28,6 +27,8 @@ pub enum Error<NoteId> {
     InvalidAmount,
 
     /// Unable to create a new spend because the wallet balance is not sufficient.
+    /// The first argument is the amount available, the second is the amount needed
+    /// to construct a valid transaction.
     InsufficientBalance(Amount, Amount),
 
     /// Chain validation detected an error in the block at the specified block height.
@@ -60,6 +61,9 @@ pub enum Error<NoteId> {
     /// The wallet attempted a sapling-only operation at a block
     /// height when Sapling was not yet active.
     SaplingNotActive,
+
+    /// It is forbidden to provide a memo when constructing a transparent output.
+    MemoForbidden,
 }
 
 impl ChainInvalid {
@@ -104,6 +108,7 @@ impl<N: fmt::Display> fmt::Display for Error<N> {
             Error::Builder(e) => write!(f, "{:?}", e),
             Error::Protobuf(e) => write!(f, "{}", e),
             Error::SaplingNotActive => write!(f, "Could not determine Sapling upgrade activation height."),
+            Error::MemoForbidden => write!(f, "It is not possible to send a memo to a transparent address."),
         }
     }
 }